Revert IdentityDbContext breaking change
